+++ openssl-sys/src/cms.rs
@@ -1,29 +1,6 @@
 use libc::*;
 use *;
 
-pub enum CMS_ContentInfo {}
-
-extern "C" {
-    #[cfg(ossl101)]
-    pub fn CMS_ContentInfo_free(cms: *mut ::CMS_ContentInfo);
-}
-
-const_ptr_api! {
-    extern "C" {
-        #[cfg(ossl101)]
-        pub fn i2d_CMS_ContentInfo(a: #[const_ptr_if(ossl300)] CMS_ContentInfo, pp: *mut *mut c_uchar) -> c_int;
-    }
-}
-
-extern "C" {
-    #[cfg(ossl101)]
-    pub fn d2i_CMS_ContentInfo(
-        a: *mut *mut ::CMS_ContentInfo,
-        pp: *mut *const c_uchar,
-        length: c_long,
-    ) -> *mut ::CMS_ContentInfo;
-}
-
 #[cfg(ossl101)]
 pub const CMS_TEXT: c_uint = 0x1;
 #[cfg(ossl101)]
@@ -68,35 +45,3 @@ pub const CMS_DEBUG_DECRYPT: c_uint = 0x20000;
 pub const CMS_KEY_PARAM: c_uint = 0x40000;
 #[cfg(ossl110)]
 pub const CMS_ASCIICRLF: c_uint = 0x80000;
-
-extern "C" {
-    #[cfg(ossl101)]
-    pub fn SMIME_read_CMS(bio: *mut ::BIO, bcont: *mut *mut ::BIO) -> *mut ::CMS_ContentInfo;
-
-    #[cfg(ossl101)]
-    pub fn CMS_sign(
-        signcert: *mut ::X509,
-        pkey: *mut ::EVP_PKEY,
-        certs: *mut ::stack_st_X509,
-        data: *mut ::BIO,
-        flags: c_uint,
-    ) -> *mut ::CMS_ContentInfo;
-
-    #[cfg(ossl101)]
-    pub fn CMS_encrypt(
-        certs: *mut stack_st_X509,
-        data: *mut ::BIO,
-        cipher: *const EVP_CIPHER,
-        flags: c_uint,
-    ) -> *mut ::CMS_ContentInfo;
-
-    #[cfg(ossl101)]
-    pub fn CMS_decrypt(
-        cms: *mut ::CMS_ContentInfo,
-        pkey: *mut ::EVP_PKEY,
-        cert: *mut ::X509,
-        dcont: *mut ::BIO,
-        out: *mut ::BIO,
-        flags: c_uint,
-    ) -> c_int;
-}
