+++ openssl-sys/src/handwritten/mod.rs
@@ -1,18 +1,27 @@
+pub use handwritten::aes::*;
 pub use handwritten::asn1::*;
 pub use handwritten::bio::*;
 pub use handwritten::bn::*;
+pub use handwritten::cms::*;
 pub use handwritten::conf::*;
 pub use handwritten::crypto::*;
 pub use handwritten::dh::*;
 pub use handwritten::dsa::*;
 pub use handwritten::ec::*;
+pub use handwritten::err::*;
 pub use handwritten::evp::*;
 pub use handwritten::hmac::*;
 pub use handwritten::kdf::*;
+pub use handwritten::object::*;
 pub use handwritten::ocsp::*;
+pub use handwritten::pem::*;
+pub use handwritten::pkcs12::*;
 pub use handwritten::pkcs7::*;
+pub use handwritten::rand::*;
 pub use handwritten::rsa::*;
+pub use handwritten::safestack::*;
 pub use handwritten::sha::*;
+pub use handwritten::srtp::*;
 pub use handwritten::ssl::*;
 pub use handwritten::stack::*;
 pub use handwritten::tls1::*;
@@ -21,21 +30,30 @@ pub use handwritten::x509::*;
 pub use handwritten::x509_vfy::*;
 pub use handwritten::x509v3::*;
 
+mod aes;
 mod asn1;
 mod bio;
 mod bn;
+mod cms;
 mod conf;
 mod crypto;
 mod dh;
 mod dsa;
 mod ec;
+mod err;
 mod evp;
 mod hmac;
 mod kdf;
+mod object;
 mod ocsp;
+mod pem;
+mod pkcs12;
 mod pkcs7;
+mod rand;
 mod rsa;
+mod safestack;
 mod sha;
+mod srtp;
 mod ssl;
 mod stack;
 mod tls1;
