+++ openssl-sys/src/handwritten/rand.rs
