+++ openssl-sys/src/handwritten/safestack.rs
