+++ openssl-sys/src/lib.rs
@@ -32,14 +32,10 @@ pub use evp::*;
 #[cfg(not(feature = "bindgen"))]
 pub use handwritten::*;
 pub use obj_mac::*;
-pub use object::*;
 pub use ocsp::*;
 pub use pem::*;
-pub use pkcs12::*;
 pub use pkcs7::*;
-pub use rand::*;
 pub use rsa::*;
-pub use safestack::*;
 pub use sha::*;
 pub use srtp::*;
 pub use ssl::*;
@@ -66,14 +62,10 @@ mod evp;
 #[cfg(not(feature = "bindgen"))]
 mod handwritten;
 mod obj_mac;
-mod object;
 mod ocsp;
 mod pem;
-mod pkcs12;
 mod pkcs7;
-mod rand;
 mod rsa;
-mod safestack;
 mod sha;
 mod srtp;
 mod ssl;
